FUSETOOLS2-2173 - Provide command to generate Camel route from OpenAPI
Note that is it provided for Yaml only. JBang is not supporting all
languages.
