+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Minimal version of VS Code to run this extension is 1.82.0
 - Update Camel Quarkus Catalog from 3.5.0 to 3.6.0
+- Provide command to create a Camel route from an OpenAPI file
 
 ## 0.13.0
