+++ src/commands/NewCamelRouteCommand.ts
@@ -16,35 +16,18 @@
  */
 'use strict'
 
-import * as path from 'path';
-import * as fs from 'fs';
-import { commands, Uri, window, workspace, WorkspaceFolder } from 'vscode';
-
-import validFilename = require('valid-filename');
+import { commands, Uri } from 'vscode';
 import { CamelInitJBangTask } from '../tasks/CamelInitJBangTask';
+import { AbstractNewCamelRouteCommand } from './AbstractNewCamelRouteCommand';
 
-export interface CamelRouteDSL {
-	language: string;
-	extension: string;
-	placeHolder: string;
-}
-
-export class NewCamelRouteCommand {
+export class NewCamelRouteCommand extends AbstractNewCamelRouteCommand {
 
 	public static ID_COMMAND_CAMEL_ROUTE_JBANG_YAML = 'camel.jbang.routes.yaml';
 	public static ID_COMMAND_CAMEL_ROUTE_JBANG_JAVA = 'camel.jbang.routes.java';
 	public static ID_COMMAND_CAMEL_ROUTE_JBANG_XML = 'camel.jbang.routes.xml';
 
-	private workspaceFolder: WorkspaceFolder | undefined;
-	private camelDSL: CamelRouteDSL | undefined;
-
-	constructor(dsl: string) {
-		this.camelDSL = this.getDSL(dsl);
-		this.workspaceFolder = this.getWorkspaceFolder()
-	}
-
 	public async create(): Promise<void> {
-		const input = await this.showInputBox();
+		const input = await this.showInputBoxForFileName();
 		if(input) {
 			const fileName = this.getFullName(input, this.camelDSL.extension);
 			const filePath = this.computeFullPath(this.workspaceFolder, fileName);
@@ -53,96 +36,4 @@ export class NewCamelRouteCommand {
 			await commands.executeCommand('vscode.open', Uri.file(filePath));
 		}
 	}
-
-	private getDSL(dsl: string): CamelRouteDSL | undefined {
-		switch (dsl) {
-			case 'YAML':
-				return { language: 'Yaml', extension: 'camel.yaml', placeHolder: 'sample-route' };
-			case 'JAVA':
-				return { language: 'Java', extension: 'java', placeHolder: 'SampleRoute' };
-			case 'XML':
-				return { language: 'Xml', extension: 'camel.xml', placeHolder: 'sample-route' };
-			default:
-				return undefined;
-		}
-	}
-
-	private async showInputBox(): Promise<string> {
-		return await window.showInputBox({
-			prompt: 'Please provide a name for the new file (without extension).',
-			placeHolder: this.camelDSL.placeHolder,
-			validateInput: (fileName) => {
-				return this.validateCamelFileName(fileName);
-			},
-		});
-	}
-
-	/**
-	 * Camel file name validation
-	 * 	- no empty name
-	 *  - name without extension
-	 *  - file already exists check
-	 *  - name cannot contains eg. special characters
-	 *  - Java pattern naming convention \b[A-Z][a-zA-Z_$0-9]*
-	 *
-	 * @param name
-	 * @returns string | undefined
-	 */
-	public validateCamelFileName(name: string): string | undefined {
-		if (!name) {
-			return 'Please provide a name for the new file (without extension).';
-		}
-		if(name.includes('.')) {
-			return 'Please provide a name without the extension.';
-		}
-		const newFilePotentialFullPath: string = this.computeFullPath(this.workspaceFolder, this.getFullName(name, this.camelDSL.extension));
-		if (fs.existsSync(newFilePotentialFullPath)) {
-			return 'The file already exists. Please choose a different file name.';
-		}
-		if (!validFilename(name)) {
-			return 'The filename is invalid.';
-		}
-		const patternJavaNamingConvention = '\\b[A-Z][a-zA-Z_$0-9]*';
-		if ((this.camelDSL.language === 'Java') && (!name.match(patternJavaNamingConvention) || name.includes(' '))) {
-			return `The filename needs to follow the ${this.camelDSL.language} naming convention. I.e. ${patternJavaNamingConvention}`;
-		}
-		return undefined;
-	}
-
-	/**
-	 * Get the full file name for provided name and suffix
-	 *
-	 * @param name of the file
-	 * @param suffix of the file
-	 * @returns the full file name format [name.suffix] eg. foo.yaml
-	 */
-	private getFullName(name: string, suffix: string): string {
-		return `${name}.${suffix}`;
-	}
-
-	/**
-	 * Resolves absolute path for the given workspace and file
-	 *
-	 * @param workspaceFolder
-	 * @param file
-	 * @returns abosolute string Path
-	 */
-	private computeFullPath(workspaceFolder: WorkspaceFolder, file: string): string {
-		return path.join(workspaceFolder.uri.fsPath, file);
-	}
-
-	/**
-	 * Resolves first opened folder in vscode existing workspace
-	 *
-	 * @returns WorkspaceFolder object
-	 */
-	private getWorkspaceFolder(): WorkspaceFolder {
-		let workspaceFolder: WorkspaceFolder | undefined;
-		if (workspace.workspaceFolders) {
-			// default to root workspace folder
-			workspaceFolder = workspace.workspaceFolders[0];
-		}
-		return workspaceFolder;
-	}
-
 }
