+++ src/extension.ts
@@ -27,6 +27,7 @@ import * as requirements from './requirements/requirements';
 import * as telemetry from './Telemetry';
 import { NewCamelQuarkusProjectCommand } from './commands/NewCamelQuarkusProjectCommand';
 import { NewCamelSpringBootProjectCommand } from './commands/NewCamelSpringBootProjectCommand';
+import { NewCamelRouteFromOpenAPICommand } from './commands/NewCamelRouteFromOpenAPICommand';
 
 const LANGUAGE_CLIENT_ID = 'LANGUAGE_ID_APACHE_CAMEL';
 const SETTINGS_TOP_LEVEL_KEY_CAMEL = 'camel';
@@ -126,6 +127,10 @@ export async function activate(context: ExtensionContext) {
 		await sendCommandTrackingEvent(NewCamelRouteCommand.ID_COMMAND_CAMEL_ROUTE_JBANG_XML);
 	}));
 
+	context.subscriptions.push(commands.registerCommand(NewCamelRouteFromOpenAPICommand.ID_COMMAND_CAMEL_ROUTE_FROM_OPEN_API_JBANG_YAML, async () => {
+		await new NewCamelRouteFromOpenAPICommand('YAML').create();
+		await sendCommandTrackingEvent(NewCamelRouteFromOpenAPICommand.ID_COMMAND_CAMEL_ROUTE_FROM_OPEN_API_JBANG_YAML);
+	}));
 
 	context.subscriptions.push(commands.registerCommand(NewCamelQuarkusProjectCommand.ID_COMMAND_CAMEL_QUARKUS_PROJECT, async () => {
 		await new NewCamelQuarkusProjectCommand().create();
