+++ src/tasks/CamelOpenAPIJBangTask.ts
@@ -1,0 +1,31 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License", destination); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+'use strict'
+
+import { TaskScope, WorkspaceFolder } from "vscode";
+import { CamelJBangTask } from "./CamelJBangTask";
+import { CamelJBang } from "../requirements/CamelJBang";
+
+export class CamelOpenAPIJBangTask extends CamelJBangTask {
+
+	constructor(scope: WorkspaceFolder | TaskScope.Workspace, routeFilename: string, openApiFilePath: string) {
+		super(scope,
+			'Init Camel Route file with JBang',
+			new CamelJBang().generateRest(routeFilename, openApiFilePath));
+	}
+
+}
